+++ lib/src/route/nav.dart
@@ -55,6 +55,7 @@ class Nav {
     bool pageOpaque = false, // 页面背景不透明
     double bgOpacityBase = 1.0,
     bool splitBottomView = false,
+    OnPreviewIndexChanged? onPreviewIndexChanged,
   }) {
     Get.to(
       () => ImagePreviewPage(
@@ -74,6 +75,7 @@ class Nav {
         imgBgOpacityBase: bgOpacityBase,
         pageBgColor: pageBgColor,
         splitBottomView: splitBottomView,
+        onPreviewIndexChanged: onPreviewIndexChanged,
       ),
       opaque: pageOpaque,
       routeName: BaseRoutes.imgPreview,